requirements.txt
click==6.7
Flask==1.0.2
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.1.0
py3exiv2==0.2.1
Werkzeug==0.14.1
